Mod/acme storage (#1)
* Add example actors for ACME Storage

* Update acme/actors/acmestoragemigrator/actor.py

Co-authored-by: Petr Stodůlka <xstodu05@gmail.com>

* Update acme/actors/acmestoragescanner/actor.py

Co-authored-by: Petr Stodůlka <xstodu05@gmail.com>

* Split acme storagescanner to two actors

The whole solution for acme contains now 3 actors, kind of
 - scanner
 - checker
 - migrator

As well, raise an error in case we cannot obtain list of active
kernel modules.

* Create new hierarhy for ISV-repos and link acme to the official one

Previously it was expected that users will link the repository using
snactor. But possibly many solutions could be created just for purpose
of inplace upgrade between specific systems (e.g. RHEL 7 -> 8) and
will not be valid for upgrade (RHEL 8 -> 9).

To make it clear, let's use the el7toel8 directory.

As the ACME repository clearly is now just for inplace upgrade from
RHEL 7 to RHEL 8, link it directly to the official
system_upgrade_el7toel8 repository.

* Update directory hierarchy in the README.md file

* ACME: run acmemigrator in the last phase and fix nitpicks

The ACME probably will not be loaded inside the initrd so nothing
can be migrated there. Instead of that, migrate the ACME storage
during the first boot.

Co-authored-by: Pavel Odvody <podvody@redhat.com>